adding first data script
