--- docs/source/api.rst
@@ -1,7 +1,0 @@
-API
-===
-
-.. autosummary::
-   :toctree: generated
-
-   ftg
