+++ docs/source/data.rst
@@ -1,0 +1,15 @@
+ftg.data
+====================
+
+.. contents:: Contents
+    :local:
+
+Data Objects
+------------
+
+.. currentmodule:: torch_geometric.data
+
+.. autosummary::
+   :nosignatures:
+   :toctree: ../generated
+   :template: autosummary/inherited_class.rst
