+++ docs/source/index.rst
@@ -16,4 +16,4 @@ Contents
 .. toctree::
 
    usage
-   api
+   data
