--- ftg.py
@@ -1,21 +1,0 @@
-"""
-fastTG - Python library for fast and efficient temporal graph learning. 
-"""
-
-__version__ = "0.1.0"
-
-class TemporalGraph(object):
-    pass
-
-
-def discretize(graph):
-    """
-    Discretize the temporal graph.
-
-    :param graph: The temporal graph object
-    :type graph: TemporalGraph or None
-    :raise ache.InvalidKindError: If the kind is invalid.
-    :return: The ingredients list.
-    :rtype: list[str]
-    """
-    return ["shells", "gorgonzola", "parsley"]
