+++ ftg/__init__.py
@@ -1,0 +1,10 @@
+"""Fast TG library for fast ML on temporal graphs."""
+
+from .data import BaseData
+
+__version__ = '0.1.0'
+
+__all__ = [
+    'BaseData',
+    '__version__',
+]
