+++ ftg/data.py
@@ -1,0 +1,45 @@
+from typing import (
+    Any,
+    Callable,
+    Dict,
+    Iterable,
+    List,
+    NamedTuple,
+    Optional,
+    Tuple,
+    Union,
+)
+
+
+class BaseData:
+    def __getattr__(self, key: str) -> Any:
+        raise NotImplementedError
+
+    def __setattr__(self, key: str, value: Any):
+        raise NotImplementedError
+
+    def __delattr__(self, key: str):
+        raise NotImplementedError
+
+    def __getitem__(self, key: str) -> Any:
+        raise NotImplementedError
+
+    def __setitem__(self, key: str, value: Any):
+        raise NotImplementedError
+
+    def __delitem__(self, key: str):
+        raise NotImplementedError
+    
+    def __copy__(self):
+        raise NotImplementedError
+
+    def __deepcopy__(self, memo):
+        raise NotImplementedError
+
+    def __repr__(self) -> str:
+        raise NotImplementedError
+
+    
+
+    
+
